Introduces models for the objects
