+++ classes/Projects.class.php
@@ -1,15 +1,4 @@
 <?php
-class Project{
-	public $id;
-	public $naam;
-	public $lead;
-	public $omschrijving;
-	public $startdatum;
-	public $sluitingsdatum;
-	public $deadline;
-	public $completed;
-	public $user;
-}
 
 class Projects{
 	private $conn = null;
