+++ classes/Tasks.class.php
@@ -3,21 +3,6 @@ require_once("classes/Database.class.php");
 require_once("classes/Users.class.php");
 require_once("classes/Projects.class.php");
 
-class Task{
-	public $id;
-	public $owner;
-	public $parent;
-	public $task_parent;
-	public $startdatum;
-	public $sluitingsdatum;
-	public $deadline;
-	public $smalltext;
-	public $bigtext;
-	public $completed;
-
-	public $project;
-	public $user;
-}
 
 class Tasks{
 	private $conn = null;
