+++ index.php
@@ -1,6 +1,9 @@
 <?php
 session_start();
 require("mail.php");
+require_once("models/Project.php");
+require_once("models/User.php");
+require_once("models/Task.php");
 require_once("classes/Tasks.class.php");
 require_once("classes/Projects.class.php");
