+++ models/Project.php
@@ -1,0 +1,12 @@
+<?php
+class Project{
+	public $id;
+	public $naam;
+	public $lead;
+	public $omschrijving;
+	public $startdatum;
+	public $sluitingsdatum;
+	public $deadline;
+	public $completed;
+	public $user;
+}
