+++ models/Task.php
@@ -1,0 +1,17 @@
+<?php
+class Task{
+	public $id;
+	public $owner;
+	public $parent;
+	public $task_parent;
+	public $startdatum;
+	public $sluitingsdatum;
+	public $deadline;
+	public $smalltext;
+	public $bigtext;
+	public $completed;
+
+	public $project;
+	public $user;
+}
+?>
